Refactored the code with some separation.
