+++ app/src/main/java/com/innovatelabs3/projectI2/domain/models/CallContent.kt
@@ -1,0 +1,6 @@
+package com.innovatelabs3.projectI2.domain.models
+
+data class CallContent(
+    val contactName: String = "",
+    val phoneNumber: String = ""
+)
