+++ app/src/main/java/com/innovatelabs3/projectI2/domain/models/ContactSaveContent.kt
@@ -1,0 +1,6 @@
+package com.innovatelabs3.projectI2.domain.models
+
+data class ContactSaveContent(
+    val name: String,
+    val phoneNumber: String
+)
