+++ app/src/main/java/com/innovatelabs3/projectI2/domain/models/GoogleMeetContent.kt
@@ -1,0 +1,5 @@
+package com.innovatelabs3.projectI2.domain.models
+
+data class GoogleMeetContent(
+    val meetingCode: String
+)
