+++ app/src/main/java/com/innovatelabs3/projectI2/domain/models/InstagramProfileContent.kt
@@ -1,0 +1,5 @@
+package com.innovatelabs3.projectI2.domain.models
+
+data class InstagramProfileContent(
+    val username: String
+)
