+++ app/src/main/java/com/innovatelabs3/projectI2/domain/models/PhonePePayment.kt
@@ -1,0 +1,7 @@
+package com.innovatelabs3.projectI2.domain.models
+
+data class PhonePePayment(
+    val amount: String,
+    val recipientUpiId: String,
+    val recipientName: String = "Recipient"
+)
