+++ app/src/main/java/com/innovatelabs3/projectI2/domain/models/ProductSearchContent.kt
@@ -1,0 +1,6 @@
+package com.innovatelabs3.projectI2.domain.models
+
+data class ProductSearchContent(
+    val query: String,
+    val platform: String = "flipkart" // "flipkart" or "amazon"
+)
