+++ app/src/main/java/com/innovatelabs3/projectI2/domain/models/SpotifySearchContent.kt
@@ -1,0 +1,6 @@
+package com.innovatelabs3.projectI2.domain.models
+
+data class SpotifySearchContent(
+    val query: String,
+    val type: String = "track" // "track" or "artist"
+)
