+++ app/src/main/java/com/innovatelabs3/projectI2/domain/models/UberRideContent.kt
@@ -1,0 +1,5 @@
+package com.innovatelabs3.projectI2.domain.models
+
+data class UberRideContent(
+    val destination: String
+)
