+++ app/src/main/java/com/innovatelabs3/projectI2/domain/models/WhatsAppMessageContent.kt
@@ -1,0 +1,7 @@
+package com.innovatelabs3.projectI2.domain.models
+
+data class WhatsAppMessageContent(
+    val contactName: String = "",
+    val phoneNumber: String = "",
+    val message: String = ""
+)
